requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
branca==0.6.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.7
crispy-bootstrap4==2022.1
cycler==0.11.0
dj-database-url==1.0.0
Django==4.1
django-admin-interface==0.26.0
django-colorfield==0.8.0
django-crispy-forms==2.0
django-heroku==0.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.39.4
frozenlist==1.3.3
geopandas==0.13.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.3.0
MarkupSafe==2.1.3
matplotlib==3.7.1
meraki==1.33.0
multidict==6.0.4
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
patsy==0.5.3
Pillow==9.2.0
plotly==5.15.0
prompt-toolkit==3.0.38
pyparsing==3.0.9
pyproj==3.6.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
requests==2.31.0
scikit-learn==0.24.2
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
sqlparse==0.4.2
statsmodels==0.14.0
tenacity==8.2.2
text-unidecode==1.3
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
xlrd==2.0.1
yarl==1.9.2
zipp==3.15.0